requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
asgiref==3.2.3
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
croniter==0.3.31
Django==3.0.3
django-scheduler==0.9.3
gevent==1.4.0
greenlet==0.4.15
helper==2.4.2
icalendar==4.0.4
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.5.0
mongoengine==0.19.1
more-itertools==8.2.0
multidict==4.7.5
numpy==1.18.1
packaging==20.1
pandas==1.0.1
pkg-resources==0.0.0
pluggy==0.13.1
polygon-api-client==0.1.1
py==1.8.1
pymongo==3.10.1
pyparsing==2.4.6
pypeln==0.1.10
pytest==5.3.5
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
schedule==0.6.0
selenium==3.141.0
six==1.14.0
sqlparse==0.3.1
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.8
websocket==0.2.1
websocket-client==0.57.0
yarl==1.4.2
zipp==3.0.0
